Start work on "Type You an Idris"; Gets TinyIdris building
Sigh... I thought getting started would at least be simple, but it turns
out a lot change in a fast-moving language like Idris in 1½ years.
The current text sets the stage and provides a patch file (along with
text explanation on the blog) for what needs changing simply for
TinyIdris (v1 and v2) to even build.