Finally write up `rotateLemma`; update wowchemy
I put this off for so long despite having the paper proof ready, due to
various other PhD things. But now it's ready!

It required maths-mode to work on the website generator, which
annoyingly (since Wowchemy apparently hates their users and don't follow
the semver stuff in the Hugo guidelines) required me to rewrite the
configs and `go.mod` file. Which in turn broke sharing. Nothing is ever
simple...