[ SPLV'20 ] Rewrite Ex2.2 based on meeting w Edwin
I finally get it, but oh boy did it take a while and is explaining it
ever difficult! I've written up the same examples that finally made it
tick for me, in the hope that it does the same for the reader. The
result is a good 2000 words extra (including code and diagrams though,
so who knows what it actually is), but also something which I am _much_
happier with!