[ admin ] Part 2/3 theme version update
gcushen must be of a similar ilk to mbostock: none of it is backwards
compatible; the documentation is scattered (actually, I think D3 is
better in that regard); some things have just...  ceased to exist??; and
others are technically in the documentation, but only if you know where
to look, what magic words to search for, and happen to know the relevant
part of the Hugo docs as well.

(ノಠ益ಠ)ノ彡┻━┻         YEEET!

I gave up in the end and just tore out the bits that had changed, copied
in the new ones from a fresh clone/zip-download, and now it builds
(minus the `coffee` theme, which might have been annihilated?).
And all it cost me was my sanity. Yaaay...