[ off-ete ] Saying goodbye to EteSync
Not a post I really wanted to write, but it's been getting increasingly
inconvenient, and I've thought about it so many times that now it was
time to just pull the trigger. I'm really hoping I'll be able to return
to EteSync one day, but for now, it's goodbye from me.